README.md
Matthew Getz Project proposal

I would like to create web application I am calling temporarily Michigan Beer Hub. I personally enjoy craft breweries and have even done some brewing with my older brother. However, with the hundreds of beers that I have tried, I do not have a concise and easily accessible place to keep track of which beers I have tried, and how much I enjoyed them. The site I propose will provide users with a login that allows them to enter Michigan beers they have had, and write comments and ratings for them. These are then kept as a log for their future use. This can also be applied to breweries that they have been to. They will be able to rate Michigan breweries to compile a list of their favorites and why. The site will have access to lists of major breweries in Michigan and the beers that they make. This will allow users to organize the beers that they enjoy and categorize the different styles of beer and which breweries that they prefer. It will also allow users to find new breweries that they might not have heard of. The site should also allow the users to look up craft breweries in Michigan and directions on how to get there. This allows users to find new breweries, visit them, rate them, rate their beers, and keep track of all that information. The reviews of breweries and beers will be available by searching the Mongo database for the comments. The site will use the MERN stack from class, and incorporate the various aspects of web development that were used in labs, and homework assignments.

I do not currently have a partner, and do not have any real preferences for who I work with. I know a handful of people in the class, but do not feel the need to work with anyone in particular.